+++ config.js
@@ -22,5 +22,8 @@ module.exports = {
 		secret: require('./config.json').jwt,
 	},
 	gcm_api_key: require('./config.json').gcm_api_key,
-	tx: require('./config.json').tx
+	tx: require('./config.json').tx,
+	domainName: require('./config.json').domainName || null,
+	dkimSelector: require('./config.json').dkimSelector || null,
+	root: __dirname
 }